requirements.txt
python-telegram-bot==20.3
Flask==3.0.3
httpx==0.27.0
python-dotenv==1.0.1
